refactor: remove most of the dependency cycles
