+++ packages/core/src/utils/DataloaderUtils.ts
@@ -3,7 +3,7 @@ import { Collection, type InitCollectionOptions } from '../entity/Collection';
 import { helper } from '../entity/wrap';
 import { type EntityManager } from '../EntityManager';
 import type DataLoader from 'dataloader';
-import { DataloaderType, ReferenceKind } from '../enums';
+import { ReferenceKind } from '../enums';
 import { type LoadReferenceOptions, Reference } from '../entity/Reference';
 
 export class DataloaderUtils {
@@ -203,17 +203,6 @@ export class DataloaderUtils {
         return entities.filter(DataloaderUtils.getColFilter(col));
       });
     };
-  }
-
-  static getDataloaderType(dataloaderCfg: DataloaderType | boolean): DataloaderType {
-    switch (dataloaderCfg) {
-      case true:
-        return DataloaderType.ALL;
-      case false:
-        return DataloaderType.NONE;
-      default:
-        return dataloaderCfg;
-    }
   }
 
 }
